plixxo_backend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
Babel==2.7.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
Django==2.2.1
django-cors-headers==3.0.2
django-enumfield==1.5
django-filter==2.1.0
django-form-utils==1.0.3
django-material==1.5.8
django-phonenumber-field==3.0.1
django-viewflow==1.5.6
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
httpie==1.0.2
idna==2.8
isort==4.3.20
kiwisolver==1.0.1
lazy-object-proxy==1.4.1
Markdown==3.1.1
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.16.1
phonenumbers==8.10.12
psycopg2==2.8.2
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.3
wrapt==1.11.1
